imporving view folder structure by spliting folders by layout
