+++ src/view/gameBoard/Cell.tsx
